+++ src/app/art/art.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { Art } from '../models';
+import { Art } from '../data/models';
 import { IonicModule } from '@ionic/angular';
 import { CommonModule } from '@angular/common';
 import { animate, state, style, transition, trigger } from '@angular/animations';
