+++ src/app/data/settings.service.ts
