+++ src/app/data/worker-interface.ts
