+++ src/app/fav-map/fav-map.page.ts
@@ -3,9 +3,9 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 import { MapComponent } from '../map/map.component';
-import { FavoritesService } from '../favorites.service';
-import { DbService } from '../db.service';
-import { MapPoint } from '../models';
+import { FavoritesService } from '../favs/favorites.service';
+import { DbService } from '../data/db.service';
+import { MapPoint } from '../data/models';
 
 @Component({
   selector: 'app-fav-map',
