+++ src/app/favs/favs.page.ts
@@ -3,14 +3,14 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonContent, IonicModule } from '@ionic/angular';
 import { Router, RouterModule } from '@angular/router';
-import { Art, Camp, Event, MapPoint } from '../models';
-import { FavoritesService } from '../favorites.service';
+import { Art, Camp, Event, MapPoint } from '../data/models';
+import { FavoritesService } from './favorites.service';
 import { EventComponent } from '../event/event.component';
-import { DbService } from '../db.service';
+import { DbService } from '../data/db.service';
 import { CampComponent } from '../camp/camp.component';
 import { MapModalComponent } from '../map-modal/map-modal.component';
 import { ArtComponent } from '../art/art.component';
-import { UiService } from '../ui.service';
+import { UiService } from '../ui/ui.service';
 import { CategoryComponent } from '../category/category.component';
 import { SearchComponent } from '../search/search.component';
 import { toMapPoint } from '../map/map.utils';
