+++ src/app/friend/friend.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { IonicModule, ModalController, ToastController } from '@ionic/angular';
-import { Friend } from '../models';
+import { Friend } from '../data/models';
 
 
 interface PickerItem {
