+++ src/app/map-modal/map-modal.component.ts
@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { MapComponent } from '../map/map.component';
 import { IonicModule } from '@ionic/angular';
-import { MapPoint } from '../models';
+import { MapPoint } from '../data/models';
 
 @Component({
   selector: 'app-map-modal',
