+++ src/app/pin-map/pin-map.page.ts
@@ -3,8 +3,8 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 import { MapComponent } from '../map/map.component';
-import { DbService } from '../db.service';
-import { MapPoint, MapSet, Pin } from '../models';
+import { DbService } from '../data/db.service';
+import { MapPoint, MapSet, Pin } from '../data/models';
 
 enum MapType {
   Restrooms = 'restrooms',
