+++ src/app/rsl/rsl.page.ts
@@ -2,12 +2,12 @@ import { Component, OnInit, ViewChild, effect } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonContent, IonicModule, ToastController } from '@ionic/angular';
-import { UiService } from '../ui.service';
-import { DbService } from '../db.service';
-import { noDate, now, sameDay } from '../utils';
-import { Day, MapPoint, RSLEvent } from '../models';
+import { UiService } from '../ui/ui.service';
+import { DbService } from '../data/db.service';
+import { noDate, now, sameDay } from '../utils/utils';
+import { Day, MapPoint, RSLEvent } from '../data/models';
 import { GpsCoord } from '../map/geo.utils';
-import { GeoService } from '../geo.service';
+import { GeoService } from '../geolocation/geo.service';
 import { SearchComponent } from '../search/search.component';
 import { MapModalComponent } from '../map-modal/map-modal.component';
 import { SkeletonEventComponent } from '../skeleton-event/skeleton-event.component';
