--- src/app/share.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ShareService } from './share.service';
-
-describe('ShareService', () => {
-  let service: ShareService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ShareService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
