+++ src/app/utils/utils.ts
@@ -1,5 +1,5 @@
 import { environment } from "src/environments/environment";
-import { OccurrenceSet, TimeString } from "./models";
+import { OccurrenceSet, TimeString } from "../data/models";
 
 export function sameDay(d1: Date, d2: Date) {
     return d1.getDate() === d2.getDate() && d1.getMonth() === d2.getMonth() && d1.getFullYear() === d2.getFullYear();
